Cleanup Chinese code (#6436)
Currently almost all the code in `chinese_based.rs` is generic in `CB:
ChineseBased`, so it will all be monomorhized twice. This reduces the
generic to just a handful of methods.

It also changes `CyclicYear::year`'s type from `NonZeroU8` to `u8`, as
we don't use `NonZero` types for any other of our 1-based outputs.

+++ components/datetime/src/error.rs
@@ -80,9 +80,9 @@ pub enum DateTimeWriteError {
     #[displaydoc("Invalid cyclic year {value} (maximum {max})")]
     InvalidCyclicYear {
         /// Value
-        value: usize,
+        value: u8,
         /// Max
-        max: usize,
+        max: u8,
     },
 
     /// The [`DecimalFormatter`] has not been loaded.
